Separate authenticate models into individual files
